trying to reorganize things a bit better
